# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: FirstMove API
  description: Activation engine API — from screen to street
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1 base path
...
paths:
...
  /activations/check:
    post:
      tags: [activations]
      summary: Trigger activation check
      description: Runs agent pipeline, returns opportunity or null
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationCheckRequest"
      responses:
        "200":
          description: Activation check result
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ActivationCardResponse"
                  - $ref: "#/components/schemas/NoOpportunityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /activations/current:
    get:
      tags: [activations]
      summary: Get current pending activation
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current activation or null
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ActivationCardResponse"
                  - type: object
                    properties:
                      activation_id: { type: "null" }
                      opportunity: { type: "null" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /activations/{id}/respond:
    post:
      tags: [activations]
      summary: Accept, dismiss, or snooze an activation
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationRespondRequest"
      responses:
        "200":
          description: Response recorded
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /activations/history:
    get:
      tags: [activations]
      summary: Past activations with outcomes
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Activation history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationHistoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # --- EVENTS ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Missing or invalid JWT
    NotFound:
      description: Resource not found
  schemas:
...
    ActivationStage:
      type: string
      enum: [suggest, recommend, precommit, activate]

    # --- Activations ---
    ActivationCheckRequest:
      type: object
      description: Optional. Location can be sent here or use last known.
      properties:
        lat: { type: number }
        lng: { type: number }
    ActivationCardResponse:
      type: object
      required: [activation_id, opportunity, stage, expires_at]
      properties:
        activation_id: { type: string, format: uuid }
        opportunity:
          $ref: "#/components/schemas/Opportunity"
        stage: { $ref: "#/components/schemas/ActivationStage" }
        expires_at: { type: string, format: date-time }
    NoOpportunityResponse:
      type: object
      properties:
        activation_id: { type: "null" }
        opportunity: { type: "null" }
        message: { type: string, example: "No suitable opportunity right now" }
    Opportunity:
      type: object
      required: [title, body, tier, walk_minutes, travel_description, starts_at, leave_by, cost_pence, venue]
      properties:
        title: { type: string }
        body: { type: string }
        tier: { $ref: "#/components/schemas/OpportunityTier" }
        walk_minutes: { type: integer }
        travel_description: { type: string }
        starts_at: { type: string, format: date-time }
        leave_by: { type: string, format: date-time }
        cost_pence: { type: integer }
        venue:
          $ref: "#/components/schemas/VenueSummary"
        social_proof:
          $ref: "#/components/schemas/SocialProof"
        commitment_action:
          $ref: "#/components/schemas/CommitmentAction"
        route_polyline: { type: string, description: "Encoded polyline for map" }
    OpportunityTier:
      type: string
      enum: [structured, recurring_pattern, micro_coordination, solo_nudge]
    VenueSummary:
      type: object
      properties:
        name: { type: string }
        lat: { type: number }
        lng: { type: number }
        address: { type: string }
    SocialProof:
      type: object
      properties:
        text: { type: string, example: "4 others going solo" }
        total_expected: { type: integer }
        solo_count: { type: integer }
        familiar_face: { type: boolean }
    CommitmentAction:
      type: object
      properties:
        type:
          type: string
          enum: [one_tap_rsvp, deep_link, internal_going, none]
        label: { type: string, example: "Hold my spot" }
        url: { type: string, format: uri, description: "For deep_link type" }
    ActivationRespondRequest:
      type: object
      required: [response]
      properties:
        response:
          type: string
          enum: [accepted, dismissed, expired, snoozed]
...
    ActivationHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ActivationHistoryItem"
        total: { type: integer }
    ActivationHistoryItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        opportunity: { $ref: "#/components/schemas/Opportunity" }
        shown_at: { type: string, format: date-time }
        response: { type: string, enum: [accepted, dismissed, expired, snoozed] }
        attended: { type: boolean }
        rating: { type: integer }
        feedback_text: { type: string }

    # --- Events ---
...